Make changes in test driver

+++ segment_trees/segment_tree.cpp
@@ -90,19 +90,19 @@ uint64_t segment_tree::update_r(uint32_t arr_start_idx, uint32_t arr_end_idx, ui
 	{
 		assert(arr_start_idx == index);
 
-		int32_t diff = value - seg_tree[current_idx];
+		int64_t diff = value - seg_tree[current_idx];
 		seg_tree[current_idx] = value;
 		return diff;
 	}
 	else if (index <= middle) //update index falls on the left tree
 	{
-		uint32_t diff = update_r(arr_start_idx, middle, 2 * current_idx + 1, index, value);
+		int64_t diff = update_r(arr_start_idx, middle, 2 * current_idx + 1, index, value);
 		seg_tree[current_idx] += diff;
 		return diff;
 	}
 	else if (index > middle) //update index falls on the right tree
 	{
-		uint32_t diff = update_r(middle + 1, arr_end_idx, 2 * current_idx + 2, index, value);
+		int64_t diff = update_r(middle + 1, arr_end_idx, 2 * current_idx + 2, index, value);
 		seg_tree[current_idx] += diff;
 		return diff;
 	}
